+++ packages/test/src/utilities/annotations.test.ts
@@ -1,9 +1,7 @@
-import { AccumulatingSimplisticHandler, deserializeLanguagesWithHandler, lioncore } from "@lionweb/core"
 import { currentSerializationFormatVersion, LionWebJsonChunk } from "@lionweb/json"
-import { genericAsTreeText, languageAsText, readFileAsJson, withoutAnnotations } from "@lionweb/utilities"
-import { readFileSync } from "fs"
+import { withoutAnnotations } from "@lionweb/utilities"
 
-import { deepEqual, equal } from "../test-utils/assertions.js"
+import { deepEqual } from "../test-utils/assertions.js"
 
 describe("annotation remover", () => {
     it("removes annotation instances and references to it", () => {
@@ -116,41 +114,3 @@ describe("annotation remover", () => {
     })
 })
 
-describe("deserializing a meta-circular language", () => {
-    it("works but reports problems", () => {
-        const serializationChunk = readFileAsJson("src/languages/io.lionweb.mps.specific.json") as LionWebJsonChunk
-        equal(
-            genericAsTreeText(serializationChunk, [lioncore]),
-            readFileSync("src/languages/io.lionweb.mps.specific.generic.txt", { encoding: "utf8" })
-        )
-
-        const gatherer = new AccumulatingSimplisticHandler()
-        const languages = deserializeLanguagesWithHandler(serializationChunk, gatherer, lioncore)
-        deepEqual(gatherer.allProblems, [
-            "can't deserialize node with id=ShortDescription-ConceptDescription: can't find the classifier with key ConceptDescription in language (io-lionweb-mps-specific, 0)",
-            "can't deserialize node with id=VirtualPackage-ConceptDescription: can't find the classifier with key ConceptDescription in language (io-lionweb-mps-specific, 0)"
-            // The deserializer is not aware that this language contains instances of annotations defined in the language itself.
-        ])
-        equal(languages.length, 1)
-        equal(languageAsText(languages[0]), readFileSync("src/languages/io.lionweb.mps.specific.m2.txt", { encoding: "utf8" }))
-    })
-
-    it("works without reporting problems after removing annotations", () => {
-        const serializationChunk = readFileAsJson("src/languages/io.lionweb.mps.specific.json") as LionWebJsonChunk
-        const preGatherer = new AccumulatingSimplisticHandler()
-        const preAnnotationLanguage = deserializeLanguagesWithHandler(withoutAnnotations(serializationChunk), preGatherer, lioncore)[0]
-        deepEqual(preGatherer.allProblems, [])
-
-        const postGatherer = new AccumulatingSimplisticHandler()
-        // (just run, don't check what comes out:)
-        deserializeLanguagesWithHandler(serializationChunk, postGatherer, lioncore, preAnnotationLanguage)
-        deepEqual(postGatherer.allProblems, [
-            "error occurred during instantiation of a node for classifier ConceptDescription with meta-pointer (io-lionweb-mps-specific, 0, ConceptDescription); reason:",
-            "Error: don't know a node of concept io.lionweb.mps.specific.ConceptDescription with key ConceptDescription that's not in LionCore M3",
-            "error occurred during instantiation of a node for classifier ConceptDescription with meta-pointer (io-lionweb-mps-specific, 0, ConceptDescription); reason:",
-            "Error: don't know a node of concept io.lionweb.mps.specific.ConceptDescription with key ConceptDescription that's not in LionCore M3"
-            // The deserializer is now aware of the existence of annotations (and knows its language), but can't instantiate instances of those,
-            // because the lioncoreWriter doesn't provide a runtime representation.
-        ])
-    })
-})